requirements.txt
streamlit
requests
python-dotenv
torch
transformers
accelerate
peft
bitsandbytes